+++ cmd/dosa/query.go
@@ -45,7 +45,7 @@ func newQueryObj(fieldName, op, valueStr string) *queryObj {
 
 // QueryOptions contains configuration for query command flags
 type QueryOptions struct {
-	Fields string `short:"f" long:"fields" description:"fields of results to return, separated by comma."`
+	Fields string `short:"f" long:"fields" description:"Fields of results to return, separated by comma."`
 }
 
 // QueryCmd is a placeholder for all query commands
@@ -54,7 +54,7 @@ type QueryCmd struct {
 	Scope         scopeFlag `short:"s" long:"scope" description:"Storage scope for the given operation." required:"true"`
 	NamePrefix    string    `short:"n" long:"namePrefix" description:"Name prefix for schema types."`
 	Prefix        string    `short:"p" long:"prefix" description:"Name prefix for schema types." hidden:"true"`
-	Path          string    `long:"path" description:"path to source." required:"true"`
+	Path          string    `long:"path" description:"Path to source." required:"true"`
 	provideClient queryClientProvider
 }
 
@@ -108,8 +108,8 @@ func (c *QueryCmd) doQueryOp(f func(ShellQueryClient, context.Context, []*queryO
 type QueryRead struct {
 	*QueryCmd
 	Args struct {
-		EntityName string   `positional-arg-name:"entity" description:"entity name"`
-		Queries    []string `positional-arg-name:"queries" description:"queries should be in the form field:operator:value, supported operator: eq"`
+		EntityName string   `positional-arg-name:"entity" description:"Entity name."`
+		Queries    []string `positional-arg-name:"queries" description:"Queries should be in the form field:operator:value, supported operator: eq."`
 	} `positional-args:"yes"`
 }
 
@@ -129,10 +129,10 @@ func (c *QueryRead) Execute(args []string) error {
 // QueryRange holds the options for 'query range'
 type QueryRange struct {
 	*QueryCmd
-	Limit int `short:"l" long:"limit" default:"100" description:"max number of results to return"`
+	Limit int `short:"l" long:"limit" default:"100" description:"Max number of results to return."`
 	Args  struct {
-		EntityName string   `positional-arg-name:"entity" description:"entity name"`
-		Queries    []string `positional-arg-name:"queries" description:"queries should be in the form field:operator:value, supported operators: eq,lt,le,gt,ge"`
+		EntityName string   `positional-arg-name:"entity" description:"Entity name."`
+		Queries    []string `positional-arg-name:"queries" description:"Queries should be in the form field:operator:value, supported operators: eq,lt,le,gt,ge."`
 	} `positional-args:"yes"`
 }