+++ CHANGELOG.md
@@ -3,6 +3,7 @@
  - Add functionality to not invalidate cache when upsert (#341)
  - Add query (read/range) support to CLI (#342)
  - Fix CLI help message (#346)
+ - CLI help messages should all start with an uppercase character (#347)
 
 ## v3.0.1 (2018-06-07)
 - Fixed version const